Add RK3588 PCIe driver implementation and update Clippy crates list
- Introduced a new driver for the Rockchip RK3588 PCIe host, implementing necessary functionalities for PCIe configuration and memory window programming.
- Added GPIO reset control for the PCIe host, allowing for proper initialization and reset handling.
- Implemented functions for mapping MMIO regions, configuring PCIe memory windows, and logging endpoint information.
- Updated the Clippy crates list to include the new `rk3588-pci` crate for linting and analysis.
