+++ Cargo.toml
@@ -108,6 +108,7 @@ members = [
     "components/x86_vcpu",
     "components/x86_vlapic",
     "drivers/npu/rockchip-npu",
+    "drivers/pci/rk3588-pci",
     "drivers/soc/rockchip/rockchip-pm",
     "drivers/soc/rockchip/rockchip-soc",
     "examples/tgmath",
@@ -303,6 +304,7 @@ range-alloc-arceos = { version = "0.3.8", path = "components/range-alloc-arceos"
 riscv-h = { version = "0.4.6", path = "components/riscv-h" }
 riscv_vcpu = { version = "0.5.6", path = "components/riscv_vcpu" }
 riscv_vplic = { version = "0.4.8", path = "components/riscv_vplic" }
+rk3588-pci = { version = "0.1.0", path = "drivers/pci/rk3588-pci" }
 rockchip-npu = { version = "0.1.1", path = "drivers/npu/rockchip-npu" }
 rockchip-pm = { version = "0.4.2", path = "drivers/soc/rockchip/rockchip-pm" }
 rockchip-soc = { version = "0.1.1", path = "drivers/soc/rockchip/rockchip-soc" }
@@ -335,6 +337,8 @@ event-listener = { version = "5.4.0", default-features = false }
 schemars = "1.2.1"
 toml = "1"
 dma-api = { version = "0.7.2" }
+mmio-api = "0.2.1"
+rdif-pcie = "0.2"
 lazy_static = { version = "1.5", features = ["spin_no_std"] }
 lock_api = { version = "0.4", default-features = false }
 log = "0.4"
